feat: Move Autoplay button to Queue
It makes more sense to be in the Queue Field rather then being in the RoomInfo Field
